+++ PSP/SpringApiRest/src/main/java/org/example/springdemo/domain/model/Usuario.java
@@ -1,10 +1,10 @@
 package org.example.springdemo.domain.model;
 
 public record Usuario(
-    Long id,
-    String username,
-    String email,
-    String nombre,
-    String rol
+        int id,
+        String username,
+        String email,
+        String nombre,
+        String rol
 ) {
 }
