+++ PSP/SpringApiRest/src/main/java/org/example/springdemo/ui/dto/EjercicioDTO.java
@@ -1,8 +1,8 @@
 package org.example.springdemo.ui.dto;
 
 public record EjercicioDTO(
-        Long id,
-        Long entrenamientoId,
+        int id,
+        int entrenamientoId,
         String nombre,
         Integer repeticiones,
         Integer series
